+++ lib/open890_web/components/slider.ex
@@ -3,6 +3,7 @@ defmodule Open890Web.Components.Slider do
 
   def slider(assigns) do
     labeled = !is_nil(assigns[:label]) && assigns[:label] != ""
+    max_value = Map.get(assigns, :max_value, 255)
 
     ~H"""
       <div class={component_classes(assigns)}>
@@ -11,7 +12,7 @@ defmodule Open890Web.Components.Slider do
         <% end %>
         <div class={wrapper_class(assigns)} phx-hook="Slider" data-click-action={@click} data-wheel-action={@wheel} id={id_for(@label)}
           data-enabled={enabled_state(assigns)}>
-          <div class="indicator" style={style_attr(@value)}></div>
+          <div class="indicator" style={style_attr(@value, max_value)}></div>
         </div>
       </div>
     """
@@ -23,9 +24,15 @@ defmodule Open890Web.Components.Slider do
     |> to_string()
   end
 
-  def style_attr(value) do
-    "width: #{value}px;"
+  def style_attr(value, max_value) when is_integer(value) and is_integer(max_value) do
+    width = value / 1
+    percentage = (width / max_value)
+    width = (percentage * 255) |> round()
+
+    "width: #{width}px;"
   end
+
+  def style_attr(nil, _), do: "width: 0px;"
 
   def id_for(label) do
     "#{label}Slider"