+++ CHANGELOG.md
@@ -1,6 +1,8 @@
 # open890 Changelog
 
 ## Unreleased
+* Added RF Power slider (#109)
+* Windows: Added a 'pause' call at the end of open890.bat to make it easier to see possible errors.
 * Added user-defined markers. Keyboard mm/mr/mg/mb to create, cc/cr/cg/cb to clear. See the wiki for more information. (#116)
 
 ## 1.0.5 - 2023-04-20